app/requirements.txt
arrow==0.15.2
pytz==2019.3
requests==2.22.0
typing==3.7.4.1
pytest==5.3.2
pytest-icdiff==0.5
pytest-instafail==0.4.2
pytest-celery==0.0.0a1
mockito==1.2.2
responses==0.12.0
coverage==5.1
Flask==1.1.1
uWSGI==2.0.18
py-healthcheck==1.9.0
celery==5.0.2
redis==3.5.3
line-bot-sdk==1.17.0
youtube-dl==2020.11.26
validators==0.18.1